feat: 모든 id 를 Long 에서 정수형으로 변경

+++ src/main/java/com/example/likelionfinalproject/domain/dto/PostResponse.java
@@ -13,7 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 public class PostResponse {
 
     private String message;
-    private Long postId;
+    private Integer postId;
 
     public static PostResponse of(Post savedPost) {
