+++ src/main/java/com/example/likelionfinalproject/repository/PostRepository.java
@@ -3,5 +3,5 @@ package com.example.likelionfinalproject.repository;
 import com.example.likelionfinalproject.domain.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Integer> {
 }
